Updates to particle filter
