+++ test/runtests.jl
@@ -20,8 +20,8 @@ include("static_dsl.jl")
 include("choice_at.jl")
 include("call_at.jl")
 include("injective.jl")
-include("inference.jl")
+include("inference/inference.jl")
 include("distribution.jl")
 include("map.jl")
-#include("unfold.jl")
+include("unfold.jl")
 #include("recurse.jl")
